Adding harness for running Mutect algorithm on real reads.
