+++ avocado-core/src/main/scala/org/bdgenomics/avocado/genotyping/BiallelicGenotyper.scala
@@ -18,18 +18,9 @@
 package org.bdgenomics.avocado.genotyping
 
 import org.apache.commons.configuration.{ HierarchicalConfiguration, SubnodeConfiguration }
-import org.apache.spark.SparkContext._
 import org.apache.spark.Logging
-import org.apache.spark.rdd.{ InstrumentedRDD, RDD }
-import org.bdgenomics.adam.models.{
-  ReferencePosition,
-  ReferenceRegion,
-  SequenceDictionary,
-  SequenceRecord,
-  VariantContext
-}
+import org.bdgenomics.adam.models._
 import org.bdgenomics.adam.rdd.ADAMContext._
-import org.bdgenomics.adam.rdd.GenomicPositionPartitioner
 import org.bdgenomics.adam.util.PhredUtils
 import org.bdgenomics.avocado.Timers._
 import org.bdgenomics.avocado.algorithms.em.EMForAlleles
@@ -38,7 +29,7 @@ import org.bdgenomics.avocado.genotyping.annotators.VariantCallingAnnotator
 import org.bdgenomics.avocado.models.{ AlleleObservation, Observation }
 import org.bdgenomics.avocado.stats.AvocadoConfigAndStats
 import org.bdgenomics.avocado.algorithms.math.LogToPhred.log2phred
-import org.bdgenomics.formats.avro.{ Contig, Genotype, GenotypeAllele, Variant, VariantCallingAnnotations }
+import org.bdgenomics.formats.avro._
 import scala.annotation.tailrec
 import scala.math.{ exp, expm1, log => mathLog, log1p, max, min, pow }
 
@@ -97,14 +88,14 @@ object BiallelicGenotyper extends GenotyperCompanion {
 }
 
 class BiallelicGenotyper(sd: SequenceDictionary,
-                         contigLengths: Map[String, Long],
+                         val contigLengths: Map[String, Long],
                          ploidy: Int = 2,
                          useEM: Boolean = false,
                          emitGVCF: Boolean = true,
                          estimatedReferenceFrequency: Double = 0.999,
                          maxIterations: Option[Int] = Some(10),
                          tolerance: Option[Double] = Some(1e-3),
-                         saturationThreshold: Double = 0.001) extends Genotyper with Logging {
+                         saturationThreshold: Double = 0.001) extends SiteGenotyper with Logging {
 
   val companion: GenotyperCompanion = BiallelicGenotyper
 
@@ -282,7 +273,7 @@ class BiallelicGenotyper(sd: SequenceDictionary,
 
   def genotypeSite(region: ReferenceRegion,
                    referenceObservation: Observation,
-                   alleleObservations: Iterable[AlleleObservation]): VariantContext = GenotypingSite.time {
+                   alleleObservations: Iterable[AlleleObservation]): Option[VariantContext] = GenotypingSite.time {
 
     // get reference allele
     val reference = referenceObservation.allele
@@ -349,126 +340,6 @@ class BiallelicGenotyper(sd: SequenceDictionary,
     })
 
     // emit the variant context
-    VariantContext(variant, genotypes, None)
-  }
-
-  def genotype(observations: RDD[Observation]): RDD[VariantContext] = {
-    val sortedRdd = observations.keyBy(_.pos)
-      .repartitionAndSortWithinPartitions(GenomicPositionPartitioner(observations.partitions.size,
-        contigLengths))
-    val instrumentedRdd = {
-      import org.apache.spark.rdd.MetricsContext._
-      sortedRdd.instrument
-    }
-    instrumentedRdd.mapPartitions(genotypeIterator)
-  }
-
-  def updateObservations(region: ReferenceRegion,
-                         obs: List[Observation]): (Observation, Iterable[AlleleObservation]) = {
-    // extract references
-    val ref = obs.flatMap(o => o match {
-      case ao: AlleleObservation => None
-      case oo: Observation       => Some(oo)
-    }).sortBy(_.pos)
-      .map(_.allele)
-      .mkString
-    val refLen = ref.length
-    val refPos = ReferencePosition(region.referenceName, region.start)
-
-    // put read observations together
-    val readObservations = obs.flatMap(o => o match {
-      case ao: AlleleObservation => Some(ao)
-      case oo: Observation       => None
-    })
-
-    // if the reference length is greater than 1, we need to expand the values
-    val finalObservations = if (refLen > 1) {
-      readObservations.groupBy(_.readId)
-        .map(ro => {
-          // sort the read observations
-          val sortedRo = ro._2.sortBy(_.pos)
-
-          // get the start and end position
-          val start = sortedRo.head.pos.pos
-          val end = sortedRo.last.pos.pos + sortedRo.last.length
-
-          // reduce down to get the sequence
-          val sequence = (ref.take((start - region.start).toInt) +
-            sortedRo.map(_.allele).mkString +
-            ref.takeRight((region.end - end).toInt))
-          val phred = sortedRo.map(_.phred).sum / sortedRo.length
-
-          // make observation
-          AlleleObservation(refPos,
-            refLen,
-            sequence,
-            phred,
-            sortedRo.head.mapq,
-            sortedRo.head.onNegativeStrand,
-            sortedRo.head.sample,
-            sortedRo.head.readId)
-        })
-    } else {
-      readObservations
-    }
-
-    // emit reference observation
-    (new Observation(refPos, ref), finalObservations)
-  }
-
-  def genotypeIterator(iter: Iterator[(ReferencePosition, Observation)]): Iterator[VariantContext] = {
-    // do we have any elements in this iterator?
-    if (iter.hasNext) {
-      // peek at the head element
-      val (headPos, headObs) = iter.next
-
-      // and initialize state
-      var windowStart = headPos.pos
-      var windowEnd = windowStart + headObs.length
-      var obsList = List(headObs)
-
-      def flush(): VariantContext = {
-        val region = ReferenceRegion(headPos.referenceName, windowStart, windowEnd)
-
-        // update observations
-        val (refObs, obsIterable) = updateObservations(region, obsList)
-
-        // compute genotypes
-        genotypeSite(region,
-          refObs,
-          obsIterable)
-      }
-
-      // start processing observation windows!
-      iter.flatMap(kv => {
-        val (pos, obs) = kv
-
-        // does this observation start past the end of this window? if so, flush our observations
-        // and rebuild the window
-        if (pos.pos >= windowEnd) {
-          // flush the window
-          val vc = flush()
-
-          // reinitialize state
-          windowStart = pos.pos
-          windowEnd = windowStart + obs.length
-          obsList = List(obs)
-
-          // emit variant context
-          Some(vc)
-        } else {
-          // make possible update to window end
-          windowEnd = max(windowEnd, pos.pos + obs.length)
-
-          // append observation to list
-          obsList = obs :: obsList
-
-          // nothing to emit
-          None
-        }
-      }) ++ Iterator(flush())
-    } else {
-      Iterator()
-    }
+    Some(VariantContext(variant, genotypes, None))
   }
 }
